+++ packages/extension-base/src/services/migration-service/scripts/EnableVaraChain.ts
@@ -1,0 +1,8 @@
+// Copyright 2019-2022 @subwallet/extension-koni authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+import EnableChain from './EnableChain';
+
+export default class EnableVaraChain extends EnableChain {
+  slug = 'vara_network';
+}