+++ packages/extension-koni-ui/src/components/WalletConnect/Network/WCNetworkBase.tsx
@@ -90,6 +90,7 @@ const WCNetworkBase = styled(Component)<Props>(({ theme: { token } }: Props) => 
       },
 
       '.ant-sw-list-wrapper': {
+        overflow: 'auto',
         flexBasis: 'auto'
       },