+++ packages/extension-koni-ui/src/stores/index.ts
@@ -11,6 +11,7 @@ import SettingsReducer from './base/Settings';
 import UIViewStateReducer from './base/UIViewState';
 import BalanceReducer from './feature/Balance';
 import BondingReducer from './feature/Bonding';
+import CampaignReducer from './feature/Campaign';
 import AssetRegistryReducer from './feature/common/AssetRegistry';
 import ChainStoreReducer from './feature/common/ChainStore';
 import CrowdloanReducer from './feature/Crowdloan';
@@ -31,7 +32,8 @@ const persistConfig = {
   whitelist: [
     'settings',
     'uiViewState',
-    'staking'
+    'staking',
+    'campaign'
   ]
 };
 
@@ -45,6 +47,7 @@ const rootReducers = combineReducers({
   balance: BalanceReducer,
   bonding: BondingReducer,
   mantaPay: MantaPayReducer,
+  campaign: CampaignReducer,
 
   // common
   chainStore: ChainStoreReducer,